requirements.txt
requests
beautifulsoup4
python-docx
openai
python-dotenv
